requirements.txt
opencv-python
tqdm
openai-whisper
moviepy
customtkinter